Change guide layout to use astro instead of vue
